Don't specify a conflicting type for setField calls
On 32bit builds, jlong is `long long`, so passing in a long and claiming
that it's a jlong is wrong. This causes a build failure after the
addition of optional exception handling for QJniObject calls in
qtbase:d045e2d2b604152450a45846fc87cb53d89bf50a.

Fix that by not specifying the field type explicitly, and instead
consistently cast to jlong (using reinterpret_cast if we have a pointer,
otherwise construct the jlong in place).

This is not a problem for calls that specify e.g. jboolean and then pass
a bool, as those are the same types on all platforms.

Fixes: QTBUG-142334
Change-Id: Icbcab1b20157b2e9d002d2de8efd9fe4acef291d
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>